services/root-agent/pyproject.toml
[project]
name = "root-agent"
version = "0.1.0"
description = "Interview agent service"
requires-python = ">=3.10"
dependencies = [
    "google-adk==1.16.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    # Code formatting and linting
    "ruff>=0.8.4",
    "pre-commit>=4.0.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
